added fixs
